requirements.txt
matplotlib
pandas
scikit-learn
plotly
streamlit
umap-learn
